now use half precision rendering

+++ Sources/CanvasKit/Canvas/Canvas.metal
@@ -9,21 +9,21 @@
 using namespace metal;
 
 struct SingleTexture {
-    texture2d<float, access::read> texture;
+    texture2d<half, access::read> texture;
 };
 
 kernel void canvas_make_texture(
     device SingleTexture* textures,
     device int2 *origins,
     constant int& textureCount,
-    texture2d<float, access::write> output,
+    texture2d<half, access::write> output,
     uint2 position [[thread_position_in_grid]]
 ) {
 //    texture.write(textures[2].texture.read(position), position);
 //    return;
     
-    float4 color = float4(0); // last component empty
-    float4 alpha = float4(0);
+    half4 color = half4(0); // last component empty
+    half4 alpha = half4(0);
     
     for (int i = 0; i < textureCount; i++) {
         int2 target_position = int2(position) - origins[i];
@@ -31,26 +31,26 @@ kernel void canvas_make_texture(
         if (target_position.x < 0 || target_position.y < 0 || target_position.x > int(textures[i].texture.get_width()) || target_position.y > int(textures[i].texture.get_height()))
             continue;
         
-        float4 newColor = textures[i].texture.read(uint2(target_position));
-        float newAlpha = newColor[3];
+        half4 newColor = textures[i].texture.read(uint2(target_position));
+        half newAlpha = newColor[3];
         
         for (int c = 0; c < 3; c++) {
-            float oldAlpha = alpha[c];
+            half oldAlpha = alpha[c];
             
-            float newComponent = newColor[c];
-            float oldComponent = color[c];
+            half newComponent = newColor[c];
+            half oldComponent = color[c];
             
-            float resultAlpha = newAlpha + oldAlpha * (1 - newAlpha);
+            half resultAlpha = newAlpha + oldAlpha * (1 - newAlpha);
             if (resultAlpha == 0) continue;
             
-            float resultComponent = (newComponent * newAlpha + oldComponent * oldAlpha * (1 - newAlpha)) / resultAlpha;
+            half resultComponent = (newComponent * newAlpha + oldComponent * oldAlpha * (1 - newAlpha)) / resultAlpha;
             
             color[c] = resultComponent;
             alpha[c] = resultAlpha;
         }
     }
     
-    float resultAlpha = max(alpha[0], max(alpha[1], alpha[2]));
+    half resultAlpha = max(alpha[0], max(alpha[1], alpha[2]));
     color[3] = resultAlpha;
     
     output.write(color, position);
