+++ Sources/CanvasKit/Layer/Selection/Select + Visible.swift
@@ -10,7 +10,7 @@ import MetalManager
 
 public struct SelectByVisible: SelectionCriteria {
     
-    let tolerance: Float
+    let tolerance: Float16
     
     public func select(layer: Layer) async throws -> Mask {
         let texture = Mask.makeTexture(width: layer.width, height: layer.height)
@@ -32,7 +32,7 @@ public extension SelectionCriteria where Self == SelectByVisible {
     /// The tolerance of 0 would only select pixels with alpha 255.
     ///
     /// The default tolerance is 254, ie, all pixels with non-zero alpha would be selected.
-    static func visible(tolerance: Float = 0.95) -> SelectionCriteria {
+    static func visible(tolerance: Float16 = 0.95) -> SelectionCriteria {
         SelectByVisible(tolerance: tolerance)
     }
